+++ feedback/lib/feedback.dart
@@ -5,5 +5,7 @@ export 'src/better_feedback.dart';
 export 'src/feedback_controller.dart';
 export 'src/feedback_mode.dart';
 export 'src/l18n/translation.dart';
-export 'src/theme/feedback_theme.dart' show FeedbackThemeData;
+export 'src/theme/feedback_theme.dart'
+    show FeedbackThemeData, FeedbackControlsThemeData;
 export 'src/user_feedback.dart';
+export 'src/controls_column.dart' show ControlsColumnBuilder;